feat: use flyway for migrations

+++ src/main/java/me/gabcytn/srsly/DTO/Problem/ReviewDetail.java
@@ -4,4 +4,4 @@ import java.time.LocalDate;
 
 import me.gabcytn.srsly.DTO.ProblemStatus;
 
-public record ReviewDetail(Integer reviewProblemId, LocalDate lastAttemptAt, LocalDate nextAttemptAt, ProblemStatus status) {}
+public record ReviewDetail(Long reviewProblemId, LocalDate lastAttemptAt, LocalDate nextAttemptAt, ProblemStatus status) {}
